+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master (unreleased)
 
 * Ignore `Style/IndentationConsistency` RuboCop warnings
+* Add `ConsecutiveControlStatements` which recommends condensing multiple
+  control statements into a single `ruby:` filter
 
 ## 0.1.0